+++ rapdev_servicenow/datadog_checks/rapdev_servicenow/constants.py
@@ -4,6 +4,8 @@ class Constants:
     ERROR_AT_LEAST_ONE_CHECK = "At least one check must be configured"
     TABLE_API_BASE_PATH = "/api/now/table"
     STATSDO_PATH = "/stats.do"
+    AGG_API_BASE_PATH = "/api/now/stats"
+    YEAR_SLUG = "/incident?sysparm_query=sys_created_onONThis%20year%40javascript%3Ags.beginningOfThisYear()%40javascript%3Ags.endOfThisYear()&sysparm_count=true"
     SNC_TIME_FORMAT = "%Y-%m-%d %H:%M:%S"
     TABLE_API_HEADERS = {
         "Accept": "application/json",
@@ -29,4 +31,4 @@ class Tables:
     SYS_USER = "sys_user"
     CMN_LOCATION = "cmn_location"
     TASK_SLA = "task_sla"
-    CONTRACT_SLA = "contract_sla"
+    CONTRACT_SLA = "contract_sla"